Make domain_validator a module
